Split tests. Running them as a suite.
Added verbosity to the running execution.

Fixes: #2
